split entity links
